New Trailer config + new config + locales fixed
The locales were not done the right way. 
New Trailer config added. 
New configuration added. Now you have option to disable Towing if said vehicle is not on the list.
This configuration is per trailer. So each trailer can be enabled or disabled.
Keybing are not configurable directly from the config
Disabling detach trailer for controller as it interfere and by default gta has dpad right to detach so it not necessary.
Long hours has been put into this.